+++ .travis.yml
@@ -3,5 +3,14 @@ elixir:
   - 1.2.2
 otp_release:
   - 18.2.1
+before_script:
+  - psql -c 'create database grit;' -U postgres
+  - MIX_ENV=test mix ecto.migrate
 after_script:
   - MIX_ENV=test mix coveralls.travis
+addons:
+  postgresql: "9.4"
+services:
+  - postgresql
+env:
+  - DATABASE_URL=postgres://postgres@localhost/grit
